Rename directory from `cratedb_jdbc` to `cratedb_tableau_jdbc`
