--- cratedb_jdbc/connectionBuilder.js
@@ -1,5 +1,0 @@
-(function dsbuilder(attr) {
-    var urlBuilder = "jdbc:postgresql://" + attr[connectionHelper.attributeServer] + ":" + attr[connectionHelper.attributePort] + "/" + attr[connectionHelper.attributeDatabase] + "?";
-
-    return [urlBuilder];
-})
