--- cratedb_jdbc/connectionMetadata.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<connection-metadata>
-  <database enabled='true'>
-    <field />
-  </database>
-  <schema enabled='true' />
-</connection-metadata>
