--- cratedb_jdbc/connectionResolver.tdr
@@ -1,24 +1,0 @@
-<?xml version='1.0' encoding='utf-8' ?>
-<tdr class='cratedb_jdbc'>
-	<connection-resolver>
-    <connection-builder>
-      <script file='connectionBuilder.js'/>
-    </connection-builder>
-    <connection-normalizer>
-      <required-attributes>
-        <attribute-list>
-          <attr>server</attr>
-          <attr>port</attr>
-          <attr>dbname</attr>
-          <attr>authentication</attr>
-          <attr>username</attr>
-          <attr>password</attr>
-          <attr>sslmode</attr>
-        </attribute-list>
-      </required-attributes>
-    </connection-normalizer>
-    <connection-properties>
-      <script file='connectionProperties.js'/>
-    </connection-properties>
-	</connection-resolver>
-</tdr>
