--- cratedb_jdbc/manifest.xml
@@ -1,40 +1,0 @@
-<?xml version='1.0' encoding='utf-8' ?>
-
-<connector-plugin class='cratedb_jdbc' superclass='jdbc' plugin-version='0.0.1' name='CrateDB' version='18.1' min-version-tableau='2020.4'>
-  <vendor-information>
-      <company name="Crate.io"/>
-      <support-link url="https://github.com/crate/tableau-connector"/>
-      <driver-download-link url="https://github.com/crate/tableau-connector"/>
-  </vendor-information>
-  <connection-customization class="cratedb_jdbc" enabled="true" version="10.0">
-    <vendor name="vendor"/>
-    <driver name="driver"/>
-    <customizations>
-      <customization name="CAP_SELECT_INTO" value="no"/>
-      <customization name="CAP_SELECT_TOP_INTO" value="no"/>
-      <customization name="CAP_CREATE_TEMP_TABLES" value="no"/>
-      <customization name="CAP_QUERY_BOOLEXPR_TO_INTEXPR" value="no"/>
-      <customization name="CAP_QUERY_GROUP_BY_BOOL" value="yes"/>
-      <customization name="CAP_QUERY_GROUP_BY_DEGREE" value="yes"/>
-      <customization name="CAP_QUERY_SORT_BY" value="yes"/>
-      <customization name="CAP_QUERY_SUBQUERIES" value="yes"/>
-      <customization name="CAP_QUERY_TOP_N" value="yes"/>
-      <customization name="CAP_QUERY_TOP_SAMPLE" value="yes"/>
-      <customization name="CAP_QUERY_TOP_SAMPLE_PERCENT" value="yes"/>
-      <customization name="CAP_QUERY_WHERE_FALSE_METADATA" value="yes"/>
-      <customization name="CAP_QUERY_SUBQUERIES_WITH_TOP" value="yes"/>
-      <customization name="CAP_SUPPORTS_SPLIT_FROM_LEFT" value="yes"/>
-      <customization name="CAP_SUPPORTS_SPLIT_FROM_RIGHT" value="yes"/>
-      <customization name="CAP_SUPPORTS_UNION" value="yes"/>
-      <customization name="CAP_QUERY_ALLOW_PARTIAL_AGGREGATION" value="no"/>
-      <customization name="CAP_QUERY_TIME_REQUIRES_CAST" value="yes"/>
-      <customization name="CAP_SUPPRESS_TEMP_TABLE_CHECKS" value="no"/>
-      <customization name="CAP_QUERY_HAVING_REQUIRES_GROUP_BY" value="yes"/>
-      <customization name="CAP_QUERY_INCLUDE_GROUP_BY_COLUMNS_IN_SELECT" value="yes"/>
-    </customizations>
-  </connection-customization>
-  <connection-fields file='connectionFields.xml'/>
-  <connection-metadata file='connectionMetadata.xml'/>
-  <connection-resolver file="connectionResolver.tdr"/>
-  <dialect file='dialect.tdd'/>
-</connector-plugin>
