+++ cratedb_tableau_jdbc/connectionBuilder.js
@@ -1,0 +1,5 @@
+(function dsbuilder(attr) {
+    var urlBuilder = "jdbc:postgresql://" + attr[connectionHelper.attributeServer] + ":" + attr[connectionHelper.attributePort] + "/" + attr[connectionHelper.attributeDatabase] + "?";
+
+    return [urlBuilder];
+})
