+++ cratedb_tableau_jdbc/connectionMetadata.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<connection-metadata>
+  <database enabled='true'>
+    <field />
+  </database>
+  <schema enabled='true' />
+</connection-metadata>
